# Redesigned "caar loop pre-boundary exchange", tuned for Frontier, also faster on Perlmutter GPU