--- a/components/homme/src/preqx_kokkos/cxx/HyperviscosityFunctorImpl.hpp
+++ b/components/homme/src/preqx_kokkos/cxx/HyperviscosityFunctorImpl.hpp
@@ -297,7 +297,7 @@ private:
   SphereOperators       m_sphere_ops;
 
   // Policies
-#if defined(KOKKOS__ENABLE_CUDA) && !defined(NDEBUG)
+#if defined(KOKKOS_ENABLE_CUDA) && !defined(NDEBUG)
   template<typename Tag>
   using TeamPolicyType = Kokkos::TeamPolicy<ExecSpace,Kokkos::LaunchBounds<512,1>,Tag>;
 #else