--- a/components/homme/src/share/cxx/utilities/ViewUtils.hpp
+++ b/components/homme/src/share/cxx/utilities/ViewUtils.hpp
@@ -67,6 +67,18 @@ viewAsReal(ViewType<ScalarType [DIM1][DIM2][DIM3], Properties...> v_in) {
   return ReturnView(reinterpret_cast<ReturnST*>(v_in.data()));
 }
 
+template <typename ScalarType, int DIM1, int DIM2, int DIM3, typename... Properties>
+KOKKOS_INLINE_FUNCTION
+typename
+std::enable_if<std::is_same<typename std::remove_const<ScalarType>::type,Scalar>::value,
+               Unmanaged<ViewType<RealType<ScalarType>*[DIM1][DIM2][DIM3*VECTOR_SIZE],Properties...>>
+              >::type
+viewAsReal(ViewType<ScalarType *[DIM1][DIM2][DIM3], Properties...> v_in) {
+  using ReturnST = RealType<ScalarType>;
+  using ReturnView = Unmanaged<ViewType<RealType<ScalarType>*[DIM1][DIM2][DIM3*VECTOR_SIZE],Properties...>>;
+  return ReturnView(reinterpret_cast<ReturnST*>(v_in.data()));
+}
+
 template <typename ScalarType, int DIM1, int DIM2, int DIM3, int DIM4, typename... Properties>
 KOKKOS_INLINE_FUNCTION
 typename
@@ -76,6 +88,30 @@ std::enable_if<std::is_same<typename std::remove_const<ScalarType>::type,Scalar>
 viewAsReal(ViewType<ScalarType [DIM1][DIM2][DIM3][DIM4], Properties...> v_in) {
   using ReturnST = RealType<ScalarType>;
   using ReturnView = Unmanaged<ViewType<RealType<ScalarType>[DIM1][DIM2][DIM3][DIM4*VECTOR_SIZE],Properties...>>;
+  return ReturnView(reinterpret_cast<ReturnST*>(v_in.data()));
+}
+
+template <typename ScalarType, int DIM1, int DIM2, int DIM3, int DIM4, typename... Properties>
+KOKKOS_INLINE_FUNCTION
+typename
+std::enable_if<std::is_same<typename std::remove_const<ScalarType>::type,Scalar>::value,
+               Unmanaged<ViewType<RealType<ScalarType>*[DIM1][DIM2][DIM3][DIM4*VECTOR_SIZE],Properties...>>
+              >::type
+viewAsReal(ViewType<ScalarType*[DIM1][DIM2][DIM3][DIM4], Properties...> v_in) {
+  using ReturnST = RealType<ScalarType>;
+  using ReturnView = Unmanaged<ViewType<RealType<ScalarType>*[DIM1][DIM2][DIM3][DIM4*VECTOR_SIZE],Properties...>>;
+  return ReturnView(reinterpret_cast<ReturnST*>(v_in.data()));
+}
+
+template <typename ScalarType, int DIM1, int DIM2, int DIM3, int DIM4, int DIM5, typename... Properties>
+KOKKOS_INLINE_FUNCTION
+typename
+std::enable_if<std::is_same<typename std::remove_const<ScalarType>::type,Scalar>::value,
+               Unmanaged<ViewType<RealType<ScalarType>*[DIM1][DIM2][DIM3][DIM4][DIM5*VECTOR_SIZE],Properties...>>
+              >::type
+viewAsReal(ViewType<ScalarType*[DIM1][DIM2][DIM3][DIM4][DIM5], Properties...> v_in) {
+  using ReturnST = RealType<ScalarType>;
+  using ReturnView = Unmanaged<ViewType<RealType<ScalarType>*[DIM1][DIM2][DIM3][DIM4][DIM5*VECTOR_SIZE],Properties...>>;
   return ReturnView(reinterpret_cast<ReturnST*>(v_in.data()));
 }