--- a/components/homme/src/theta-l_kokkos/cxx/CaarFunctorImpl.hpp
+++ b/components/homme/src/theta-l_kokkos/cxx/CaarFunctorImpl.hpp
@@ -30,6 +30,7 @@
 #include "profiling.hpp"
 #include "ErrorDefs.hpp"
 
+#include <algorithm>
 #include <assert.h>
 
 namespace Homme {
@@ -107,7 +108,7 @@ struct CaarFunctorImpl {
   struct TagPostExchange {};
 
   // Policies
-#ifndef NDEBUG
+#if defined(KOKKOS_ENABLE_CUDA) && !defined(NDEBUG)
   template<typename Tag>
   using TeamPolicyType = Kokkos::TeamPolicy<ExecSpace,Kokkos::LaunchBounds<512,1>,Tag>;
 #else
@@ -123,6 +124,16 @@ struct CaarFunctorImpl {
 
   Kokkos::Array<std::shared_ptr<BoundaryExchange>, NUM_TIME_LEVELS> m_bes;
 
+  template <bool HYDROSTATIC, bool CONSERVATIVE> void epoch1_blockOps(const SphereOuter::Team &);
+  template <bool RSPLIT_ZERO> void epoch2_scanOps(const SphereOuter::Team &);
+  template <bool HYDROSTATIC, bool RSPLIT_ZERO> void epoch3_blockOps(const SphereOuter::Team &);
+  void epoch4_scanOps(const SphereOuter::Team &);
+  template <bool HYDROSTATIC, bool RSPLIT_ZERO> void epoch5_colOps(const SphereOuter::Team &);
+  template <bool HYDROSTATIC, bool RSPLIT_ZERO, bool PGRAD_CORRECTION> void epoch6_blockOps(const SphereOuter::Team &);
+  template <bool HYDROSTATIC, bool RSPLIT_ZERO> void epoch7_col(const SphereOuter::Team &);
+  void caar_compute();
+
+  public:
   CaarFunctorImpl(const Elements &elements, const Tracers &/* tracers */,
                   const ReferenceElement &ref_FE, const HybridVCoord &hvcoord,
                   const SphereOperators &sphere_ops, const SimulationParams& params)
@@ -180,7 +191,7 @@ struct CaarFunctorImpl {
 
   int requested_buffer_size () const {
     // Ask the buffers manager to allocate enough buffers to satisfy Caar's needs
-    const int nslots = m_tu.get_num_ws_slots();
+    const int nslots = std::max(m_tu.get_num_ws_slots(), m_num_elems);
 
     int num_scalar_mid_buf = Buffers::num_3d_scalar_mid_buf;
     int num_scalar_int_buf = Buffers::num_3d_scalar_int_buf;
@@ -192,7 +203,7 @@ struct CaarFunctorImpl {
     if (m_theta_hydrostatic_mode) {
       // pi=pnh, and no wtens/phitens
       num_scalar_mid_buf -= 1;
-      num_scalar_int_buf -= 3;
+      num_scalar_int_buf -= 2;
 
       // No grad_w_i/v_i
       num_vector_int_buf -= 2;
@@ -200,10 +211,6 @@ struct CaarFunctorImpl {
     if (m_rsplit>0) {
       // No theta_i/eta_dot_dpdn
       num_scalar_int_buf -=2;
-      if (m_theta_hydrostatic_mode) {
-        // No dp_i
-        num_scalar_int_buf -= 1;
-      }
     }
 
     return num_scalar_mid_buf  *NP*NP*NUM_LEV  *VECTOR_SIZE*nslots
@@ -216,7 +223,7 @@ struct CaarFunctorImpl {
     Errors::runtime_check(fbm.allocated_size()>=requested_buffer_size(), "Error! Buffers size not sufficient.\n");
 
     Scalar* mem = reinterpret_cast<Scalar*>(fbm.get_memory());
-    const int nslots = m_tu.get_num_ws_slots();
+    const int nslots = std::max(m_tu.get_num_ws_slots(), m_num_elems);
 
     // Midpoints scalars
     m_buffers.pnh        = decltype(m_buffers.pnh       )(mem,nslots);
@@ -260,10 +267,8 @@ struct CaarFunctorImpl {
     mem += m_buffers.v_tens.size();
 
     // Interface scalars
-    if (!m_theta_hydrostatic_mode || m_rsplit==0) {
-      m_buffers.dp_i = decltype(m_buffers.dp_i)(mem,nslots);
-      mem += m_buffers.dp_i.size();
-    }
+    m_buffers.dp_i = decltype(m_buffers.dp_i)(mem,nslots);
+    mem += m_buffers.dp_i.size();
 
     if (!m_theta_hydrostatic_mode) {
       m_buffers.dpnh_dp_i = decltype(m_buffers.dpnh_dp_i)(mem,nslots);
@@ -280,9 +285,9 @@ struct CaarFunctorImpl {
     if (!m_theta_hydrostatic_mode) {
       m_buffers.phi_tens     = decltype(m_buffers.phi_tens    )(mem,nslots);
       mem += m_buffers.phi_tens.size();
-      m_buffers.w_tens       = decltype(m_buffers.w_tens      )(mem,nslots);
-      mem += m_buffers.w_tens.size();
     }
+    m_buffers.w_tens       = decltype(m_buffers.w_tens      )(mem,nslots);
+    mem += m_buffers.w_tens.size();
 
     // Interface vectors
     if (!m_theta_hydrostatic_mode) {
@@ -346,12 +351,23 @@ struct CaarFunctorImpl {
     profiling_resume();
 
     GPTLstart("caar compute");
+
+#if 0
+
+    caar_compute();
+    Kokkos::fence();
+    GPTLstop("caar compute");
+
+#else
+
     int nerr;
     Kokkos::parallel_reduce("caar loop pre-boundary exchange", m_policy_pre, *this, nerr);
     Kokkos::fence();
     GPTLstop("caar compute");
     if (nerr > 0)
       check_print_abort_on_bad_elems("CaarFunctorImpl::run TagPreExchange", data.n0);
+
+#endif
 
     GPTLstart("caar_bexchV");
     m_bes[data.np1]->exchange(m_geometry.m_rspheremp);
@@ -544,8 +560,8 @@ struct CaarFunctorImpl {
       auto dp      = Homme::subview(m_state.m_dp3d,kv.ie,m_data.n0,igp,jgp);
       auto div_vdp = Homme::subview(m_buffers.div_vdp,kv.team_idx,igp,jgp);
       auto pi      = Homme::subview(m_buffers.pi,kv.team_idx,igp,jgp);
-      auto omega_i = Homme::subview(m_buffers.grad_phinh_i,kv.team_idx,0,igp,jgp);
-      auto pi_i    = Homme::subview(m_buffers.grad_phinh_i,kv.team_idx,1,igp,jgp);
+      auto omega_i = Homme::subview(m_buffers.w_tens,kv.team_idx,igp,jgp);
+      auto pi_i    = Homme::subview(m_buffers.dp_i,kv.team_idx,igp,jgp);
 
       Kokkos::single(Kokkos::PerThread(kv.team),[&]() {
         pi_i(0)[0] = m_hvcoord.ps0*m_hvcoord.hybrid_ai0;
@@ -566,11 +582,13 @@ struct CaarFunctorImpl {
       kv.team_barrier();
 
       ColumnOps::column_scan_mid_to_int<true>(kv,div_vdp,omega_i);
+      kv.team_barrier();
       // Average omega_i to midpoints, and change sign, since later
       //   omega=v*grad(pi)-average(omega_i)
       auto omega = Homme::subview(m_buffers.omega_p,kv.team_idx,igp,jgp);
       ColumnOps::compute_midpoint_values<CombineMode::Scale>(kv,omega_i,omega,-1.0);
     });
+
     kv.team_barrier();
 
     // Compute grad(pi)
@@ -642,6 +660,7 @@ struct CaarFunctorImpl {
         // Compute interface horiz velocity
         auto u_i  = Homme::subview(m_buffers.v_i,kv.team_idx,0,igp,jgp);
         auto v_i  = Homme::subview(m_buffers.v_i,kv.team_idx,1,igp,jgp);
+
         ColumnOps::compute_interface_values(kv.team,dp,dp_i,u,u_i);
         ColumnOps::compute_interface_values(kv.team,dp,dp_i,v,v_i);
 
@@ -1138,7 +1157,6 @@ struct CaarFunctorImpl {
         ColumnOps::compute_midpoint_values<CombineMode::Scale>(kv, w_sq, Homme::subview(m_buffers.temp,kv.team_idx,igp,jgp),0.5);
       });
       kv.team_barrier();
-
       // Compute grad(average(w^2/2)). Store in wvor.
       m_sphere_ops.gradient_sphere(kv, Homme::subview(m_buffers.temp,kv.team_idx),
                                        wvor);
@@ -1179,8 +1197,8 @@ struct CaarFunctorImpl {
       if (!m_theta_hydrostatic_mode) {
         // Compute wvor = grad(average(w^2/2)) - average(w*grad(w))
         // Note: vtens is already storing grad(avg(w^2/2))
-        auto gradw_x = Homme::subview(m_buffers.v_i,kv.team_idx,0,igp,jgp);
-        auto gradw_y = Homme::subview(m_buffers.v_i,kv.team_idx,1,igp,jgp);
+        auto gradw_x = Homme::subview(m_buffers.grad_w_i,kv.team_idx,0,igp,jgp);
+        auto gradw_y = Homme::subview(m_buffers.grad_w_i,kv.team_idx,1,igp,jgp);
         auto w_i = Homme::subview(m_state.m_w_i,kv.ie,m_data.n0,igp,jgp);
 
         const auto w_gradw_x = [&gradw_x,&w_i] (const int ilev) -> Scalar {