--- a/components/homme/src/theta-l_kokkos/cxx/LimiterFunctor.hpp
+++ b/components/homme/src/theta-l_kokkos/cxx/LimiterFunctor.hpp
@@ -49,7 +49,7 @@ struct LimiterFunctor {
   struct TagDp3dLimiter {};
 
   // Policies
-#ifndef NDEBUG
+#if defined(KOKKOS_ENABLE_CUDA) && !defined(NDEBUG)
   template<typename Tag>
   using TeamPolicyType = Kokkos::TeamPolicy<ExecSpace,Kokkos::LaunchBounds<512,1>,Tag>;
 #else