… components/homme/cmake/machineFiles/frontier-bfb-serial.cmake
… components/homme/src/theta-l_kokkos/CMakeLists.txt
… components/homme/src/theta-l_kokkos/cxx/CaarFunctorImpl.cpp