This pull request attempts to provide Frontier optimizations even better than those used in the 2023 Gordon Bell Climate runs, but with a software architecture that meets the requirements of https://acme-climate.atlassian.net/wiki/x/oICd6, and with additional changes to reduce slowdown on Perlmutter CPUs.

It replaces pull request https://github.com/E3SM-Project/E3SM/pull/6522.

Summary of changes:
* Create multiple new `struct`s in `SphereOperators.hpp` to use for the new Caar pre-boundary exchange.
* Create macros in `SphereOperators.hpp` that allow code that uses implicit parallelism and vector registers on GPUs to add explicit loops and temporary arrays on CPUs.
* Use `#if (WARP_SIZE == 1)` preprocessor directives in `SphereOperators.hpp` to try to minimize CPU-specific code.
* Add indexing functions like `zbelow` in `SphereOperators.hpp` to support `Scalar` types with `VECTOR_SIZE` > 1 on CPUs.
* Add new file `CaarFunctorImpl.cpp` that implements the new `caar_compute` function and template functions with Kokkos loops. The single source code supports both GPUs and CPUs by relying on functions and macros defined in `SphereOperators.hpp`.
* Update `CaarFunctorImpl.hpp` with new functions, slight changes to temporary buffers, and `#if` to turn on/off the new `caar_compute`. If we adopt the new implementation permanently, significant code can be eliminated from this file.
* Add new `viewAsReal` functions in `ViewUtils.hpp`.
* Add preprocessor directives around `LaunchBounds<512,1>` calls, which I think are incorrect for AMD GPUs, where the Kokkos teams sometimes use 1024 threads.
* Add `frontier-bfb.cmake`, `frontier-bfb-serial.cmake`, and `pm-cpu-bfb.cmake` files for bit-for-bit unit testing of Caar.

I confirmed that the modified code passes the `caar_ut` unit test, and I ran a single-node NE30 test from Noel Keen on Frontier, Perlmutter GPU, and Perlmutter CPU. Here is a comparison of total "caar compute" times, summed over all MPI tasks on the node (8 on Frontier GPU, 4 on Perlmutter GPU, 128 on Perlmutter CPU).

| Machine | Original Code `#if 0` | New Code `#if 1` | Speedup |
| --- | --- | --- | --- |
| Frontier GPU | 182.5 | 73.42 | 2.49 |
| Perlmutter GPU | 85.36 | 55.69 | 1.53 |
| Perlmutter CPU Gnu | 1245 | 1370 | 0.909 |
| Perlmutter CPU Intel | 680.1 | 1295 | 0.525 |

The good news is that the new code is faster on both Frontier and Perlmutter GPUs. The bad news is that it slows down Perlmutter CPUs. In particular, it appears to inhibit whatever optimization the Intel compiler is able to do over the Gnu compiler.