@ambrad 2025-02-05
> Here is a comparison of total "caar compute" times, summed over all MPI tasks on the node

Trey, for clarity, would you mind explaining how you take a line from a timer file and derive the entry in the table?