@trey-ornl 2025-02-05
> Trey, for clarity, would you mind explaining how you take a line from a timer file and derive the entry in the table?

The times listed above are from the fourth column of numbers in the timer output, `1.360819e+04` in your example. Here is the raw output used for the table.

Frontier GPU original:
```
login03:/lustre/orion/proj-shared/cli200/trey/nk-ne30/cases/t.caar.ne30pg2_ne30pg2.F2010-SCREAMv1.frontier-scream-gpu.craygnuamdgpu.5d.nocosp.p3sk.shocsk.25-01-02-11.58.09/case_scripts/timing$ zgrep 'caar compute' e3sm_timing_stats.2921621.250108-174410.gz
"a:caar compute"                                              -          8        8 1.152000e+05   1.825385e+02    22.864 (     2      0)    22.764 (     1      0)
```
Frontier GPU new:
```
login03:/lustre/orion/proj-shared/cli200/trey/nk-ne30/cases/t.caar.ne30pg2_ne30pg2.F2010-SCREAMv1.frontier-scream-gpu.craygnuamdgpu.5d.nocosp.p3sk.shocsk.25-01-02-11.58.09/case_scripts/timing$ zgrep 'caar compute' e3sm_timing_stats.2921576.250108-171302.gz
"a:caar compute"                                              -          8        8 1.152000e+05   7.342101e+01     9.192 (     7      0)     9.147 (     0      0)
```
Perlmutter GPU original:
```
trey@perlmutter:login30:/pscratch/sd/t/trey/nk-ne30/cases/t.caar.ne30pg2_ne30pg2.F2010-SCREAMv1.pm-gpu.gnugpu.5d.nocosp.p3sk.shocsk.25-01-09-08.22.26/case_scripts/timing> zgrep 'caar compute' e3sm_timing_stats.34721245.250109-100816.gz
"a:caar compute"                                              -          4        4 5.760000e+04   8.706921e+01    21.814 (     3      0)    21.730 (     1      0)
```
Perlmutter GPU new:
```
trey@perlmutter:login30:/pscratch/sd/t/trey/nk-ne30/cases/t.caar.ne30pg2_ne30pg2.F2010-SCREAMv1.pm-gpu.gnugpu.5d.nocosp.p3sk.shocsk.25-01-09-08.22.26/case_scripts/timing> zgrep 'caar compute' e3sm_timing_stats.34719801.250109-083940.gz
"a:caar compute"                                              -          4        4 5.760000e+04   5.568991e+01    14.070 (     2      0)    13.750 (     0      0)
```
Perlmutter CPU Gnu original:
```
trey@perlmutter:login30:/pscratch/sd/t/trey/nk-ne30/cases/t.caar.ne30pg2_ne30pg2.F2010-SCREAMv1.pm-cpu.gnu.5d.nocosp.p3sk.shocsk.25-01-09-07.54.07/case_scripts/timing> zgrep 'caar compute' e3sm_timing_stats.34720749.250109-091742.gz
"a:caar compute"                                              -        128      128 1.843200e+06   1.244864e+04   102.537 (    16      0)    94.639 (   101      0)
```
Perlmutter CPU Gnu new:
```
trey@perlmutter:login30:/pscratch/sd/t/trey/nk-ne30/cases/t.caar.ne30pg2_ne30pg2.F2010-SCREAMv1.pm-cpu.gnu.5d.nocosp.p3sk.shocsk.25-01-09-07.54.07/case_scripts/timing> zgrep 'caar compute' e3sm_timing_stats.34719528.250109-083936.gz
"a:caar compute"                                              -        128      128 1.843200e+06   1.369728e+04   138.202 (    11      0)    87.969 (   101      0)
```
Perlmutter CPU Intel original:
```
trey@perlmutter:login30:/pscratch/sd/t/trey/nk-ne30/cases/t.caar.ne30pg2_ne30pg2.F2010-SCREAMv1.pm-cpu.intel.5d.nocosp.p3sk.shocsk.25-01-09-08.03.14/case_scripts/timing> zgrep 'caar compute' e3sm_timing_stats.34720964.250109-093013.gz
"a:caar compute"                                              -        128      128 1.843200e+06   6.800835e+03    61.787 (    14      0)    47.484 (   101      0)
```
Perlmutter CPU Intel new:
```
trey@perlmutter:login30:/pscratch/sd/t/trey/nk-ne30/cases/t.caar.ne30pg2_ne30pg2.F2010-SCREAMv1.pm-cpu.intel.5d.nocosp.p3sk.shocsk.25-01-09-08.03.14/case_scripts/timing> zgrep 'caar compute' e3sm_timing_stats.34719150.250109-082107.gz
"a:caar compute"                                              -        128      128 1.843200e+06   1.295275e+04   136.372 (    14      0)    80.493 (   103      0)
```