@trey-ornl 2025-02-05
> Thanks, Trey. That clears things up. Note that in your table you wrote 1.244864e+04 as 1245 rather than 12449, and similarly for the other CPU numbers.

Oops! Fixed.