@trey-ornl 2025-02-11
> Would it make sense to break this PR up? It looks like there may be changes not specifically related to the new CAAR design?

I would say that all the changes _are_ related to the new Caar design. The new BFB build files, for example, are only tested for the Caar unit test. I think each change can be tracked back to a Caar dependence.