@ndkeen 2025-02-24
Still unsure about the double underscore noted above, but I tried the NESAP ne1024 benchmark (without double underscore) on pm-gpu and we are seeing around 9% overall improvement.  Though there were a number of changes that may have also had some impact.